Make integration tests more robust and show proper error messages

+++ test_integration.sh
@@ -1,13 +1,14 @@
-#!/bin/sh
+#!/bin/bash
 
-set -e
+set -e -x
 
 NAME=${NAME:-jacekkow/pyveth}
 VERSION=${VERSION:-latest}
 
 PLUGIN="${NAME}:${VERSION}"
 
-docker plugin install jacekkow/pyipam:latest || true
+docker network rm test1 || true
+docker network rm test2 || true
 
 docker plugin install jacekkow/pyipam:latest || true
 
@@ -28,8 +29,14 @@ ADDRESSES=$(docker run --rm --network test1 \
   alpine \
   /sbin/ip addr show
 )
-echo "${ADDRESSES}" | grep 192.168.255.129/24
-echo "${ADDRESSES}" | grep 2001:db8:aaaa:bbbb::1/32
+if ! echo "${ADDRESSES}" | grep 192.168.255.129/24; then
+	echo "ERROR: invalid IPv4 address assigned"
+	exit 1
+fi
+if ! echo "${ADDRESSES}" | grep 2001:db8:aaaa:bbbb::1/32; then
+	echo "ERROR: invalid IPv6 address assigned"
+	exit 1
+fi
 
 
 ADDRESSES=$(docker run --rm --network test1 \
@@ -37,8 +44,14 @@ ADDRESSES=$(docker run --rm --network test1 \
   alpine \
   /sbin/ip addr show
 )
-echo "${ADDRESSES}" | grep 192.168.255.25/24
-echo "${ADDRESSES}" | grep 2001:db8:dddd:eeee:ffff:1:2:3/32
+if ! echo "${ADDRESSES}" | grep 192.168.255.25/24; then
+	echo "ERROR: invalid IPv4 address assigned"
+	exit 1
+fi
+if ! echo "${ADDRESSES}" | grep 2001:db8:dddd:eeee:ffff:1:2:3/32; then
+	echo "ERROR: invalid IPv6 address assigned"
+	exit 1
+fi
 
 docker network rm test1
 
@@ -57,12 +70,18 @@ ROUTES=$(docker run --rm --network test2 \
   alpine \
   /sbin/ip route show
 )
-echo "${ROUTES}" | grep 192.168.255.254
+if ! echo "${ROUTES}" | grep 192.168.255.254; then
+	echo "ERROR: invalid IPv4 route"
+	exit 1
+fi
 
 ROUTES=$(docker run --rm --network test2 \
   alpine \
   /sbin/ip -6 route show
 )
-echo "${ROUTES}" | grep 2001:db8:ffff:ffff:ffff:ffff:ffff:ffff
+if ! echo "${ROUTES}" | grep 2001:db8:ffff:ffff:ffff:ffff:ffff:ffff; then
+	echo "ERROR: invalid IPv6 route"
+	exit 1
+fi
 
 docker network rm test2
